requirements.txt
mcp==1.9.3
asyncpg==0.30.0
python-dotenv==1.1.0
sqlparse==0.5.3